Refactor test ESWorker to separate class
This doesn't technically remove a circular dependency, but it does
remove an awkward one.
